requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
scipy
opencv-python
einops
numpy
pillow
facexlib
gfpgan